requirements.txt
Flask==2.0.1
Werkzeug==2.2.2
telethon==1.38.1
python-dotenv==1.0.0
aiohttp==3.11.2